Address review comments: * Move `block_nodes.rs` to Dom. * Add comments to clarify behaviours.
