+++ crates/wysiwyg/src/composer_model/quotes.rs
@@ -32,7 +32,7 @@ where
 
     fn add_quote(&mut self) -> ComposerUpdate<S> {
         let (s, e) = self.safe_selection();
-        let Some(wrap_result) = self.find_nodes_to_wrap_in_block(s, e) else {
+        let Some(wrap_result) = self.state.dom.find_nodes_to_wrap_in_block(s, e) else {
             // No suitable nodes found to be wrapped inside the quote. The Dom should be empty
             self.state.dom.append_at_end_of_document(DomNode::new_quote(vec![DomNode::new_text(S::zwsp())]));
             self.state.start += 1;
